move scan to separate module
